Add files via upload
Created my first javaScript promise, to fetch data from php file. In the older version I used synchronous call to do that, because the arrays from the back end must be loaded prior to the execution of all functions, so Promises sorted this issue. Now the weekstart variables are loaded into an array firstly and that array can be accessed from everywhere in the file.
Also fixed the hidehourselect function which was missing else statements.