Add files via upload
Started writing the loadResponseData function.
So far I have destructured the response object (extracted values and assigned them to a variables of the same name as object parameter name).
Converted all of them into Numbers, since some of them have the type of a string, and I need to do some simple calculations using those variables, plus I intend to use toFixed() method which requires a variable to be of number type.
And lastly I picked elements from the DOM, in to which I will load the response data. (also did not pick the elements for charts, will do that some time later).