Add files via upload
Started adding data into tables. So far 5 tables get their data in to them.
done some additional styling.

ps. there has to be a way to create a number with two decimals on the back end (now it looses 0: if it's 3.50 the last 0 gets cut of), If that could be sorted, there would be very minimal use of Number and toFixed(2) methods on the front end. Need to look in to this later.