Add files via upload
Discarded the fetch method to pick up data from the server, as I was having issues with php session_start.
For the last few days was working more on the landing page, with login, lost password and register modals, and rewriting the back end for all of these pages. Transformed login, registration and password recovery into asynchronous. Now the next goal is to rewrite payment, deductions and account settings pages! Now time to rest :D