Add files via upload
Uploaded old versions of payment settings page and javascript files. Frist thing to do, is to add boostrap grid, nav, modals nad footer to the html file while maintaining all the same id. After that will need to take care of javascript, as it is not as efficient  and tidy as it could be.