Add files via upload
Added Bootstrap to deductionsettings.php and amended the JS file to work properly. Also replaced the type="text" inside input fields to type="number" where the value to be entered is expected to be a number. 
Some experimenting with styling done as well.

ps: Not only do I need to rewrite the JS files, but i need to decide how error handling should be done across all pages, because currently it is left inconsistent.